[ui] Delete old navigation, clean up observe
[INTERNAL_BRANCH=dish/remove-old-feature-gates]
