+++ js_modules/dagster-ui/packages/ui-core/src/app/AppLayout.tsx
@@ -1,31 +1,33 @@
-import {FeatureFlag} from 'shared/app/FeatureFlags.oss';
-
-import {AppTopNav} from './AppTopNav/AppTopNav';
-import {AppWithNewLayout} from './AppWithNewLayout';
-import {AppWithOldLayout} from './AppWithOldLayout';
-import {featureEnabled} from './Flags';
-import {HelpMenu} from './HelpMenu';
-import {UserSettingsButton} from './UserSettingsButton';
+import {useFullScreen} from './AppTopNav/AppTopNavContext';
+import {useJobStateForNav} from './AppTopNav/useJobStateForNav';
+import {AppContainer} from './navigation/AppContainer';
+import {NavCollapseProvider} from './navigation/NavCollapseProvider';
+import {getBottomGroups, getTopGroups} from './navigation/mainNavigationItems';
+import {useFeatureFlags} from './useFeatureFlags';
 
 interface Props {
+  banner?: React.ReactNode;
   children: React.ReactNode;
 }
-export const AppLayout = ({children}: Props) => {
-  const newLayout = featureEnabled(FeatureFlag.flagNavigationUpdate);
-  if (newLayout) {
-    return <AppWithNewLayout>{children}</AppWithNewLayout>;
-  }
+
+export const AppLayout = ({banner, children}: Props) => {
+  const featureFlags = useFeatureFlags();
+  const {isFullScreen} = useFullScreen();
+  const jobState = useJobStateForNav();
+
+  const topGroups = getTopGroups({featureFlags, jobState});
+  const bottomGroups = getBottomGroups({featureFlags, jobState});
 
   return (
-    <AppWithOldLayout
-      top={
-        <AppTopNav allowGlobalReload>
-          <HelpMenu showContactSales={false} />
-          <UserSettingsButton />
-        </AppTopNav>
-      }
-    >
-      {children}
-    </AppWithOldLayout>
+    <NavCollapseProvider>
+      <AppContainer
+        topGroups={topGroups}
+        bottomGroups={bottomGroups}
+        banner={banner}
+        isFullScreenEnabled={isFullScreen}
+      >
+        {children}
+      </AppContainer>
+    </NavCollapseProvider>
   );
 };
