+++ js_modules/dagster-ui/packages/ui-core/src/app/DefaultFeatureFlags.oss.tsx
@@ -4,7 +4,6 @@ import {FeatureFlag} from 'shared/app/FeatureFlags.oss';
  * Default values for feature flags when they are unset.
  */
 export const DEFAULT_FEATURE_FLAG_VALUES: Partial<Record<FeatureFlag, boolean>> = {
-  [FeatureFlag.flagNavigationUpdate]: true,
   [FeatureFlag.flagAssetCatalogSidebar]: true,
 
   // Flags for tests
