+++ js_modules/dagster-ui/packages/ui-core/src/app/FeatureFlags.oss.tsx
@@ -3,9 +3,7 @@ export enum FeatureFlag {
   flagDisableWebsockets = 'flagDisableWebsockets',
   flagSidebarResources = 'flagSidebarResources',
   flagDisableAutoLoadDefaults = 'flagDisableAutoLoadDefaults',
-  flagUseNewObserveUIs = 'flagUseNewObserveUIs',
   flagAssetGraphGroupsPerCodeLocation = 'flagAssetGraphGroupsPerCodeLocation',
-  flagNavigationUpdate = 'flagNavigationUpdate',
   flagAssetCatalogSidebar = 'flagAssetCatalogSidebar',
 
   // Flags for tests
