--- js_modules/dagster-ui/packages/ui-core/src/app/observeEnabled.oss.tsx
@@ -1,3 +1,0 @@
-export const observeEnabled = (): boolean => {
-  return false;
-};
