+++ js_modules/dagster-ui/packages/ui-core/src/assets/LaunchAssetExecutionButton.tsx
@@ -13,7 +13,6 @@ import pick from 'lodash/pick';
 import uniq from 'lodash/uniq';
 import React, {useContext, useMemo, useState} from 'react';
 import {Link} from 'react-router-dom';
-import {observeEnabled} from 'shared/app/observeEnabled.oss';
 import {useLaunchWithTelemetry} from 'shared/launchpad/useLaunchWithTelemetry.oss';
 
 import {ASSET_NODE_CONFIG_FRAGMENT} from './AssetConfig';
@@ -192,7 +191,7 @@ export function optionsForExecuteButton(
     materializeOption: {
       assetKeys: materializable.map((a) => a.assetKey),
       disabledReason: materializationDisabledReason(assets, materializable, isSingle),
-      icon: <Icon name={observeEnabled() ? 'execute' : 'materialization'} />,
+      icon: <Icon name="execute" />,
       label: isSelection
         ? `Materialize selected${countIfPluralOrNotAll(materializable, assets)}${ellipsis}`
         : materializable.length > 1 && !skipAllTerm
@@ -328,13 +327,7 @@ export const LaunchAssetExecutionButton = ({
                 borderRight: `1px solid rgba(255,255,255,0.2)`,
               }}
               disabled={!canLaunch}
-              icon={
-                loading ? (
-                  <Spinner purpose="body-text" />
-                ) : (
-                  <Icon name={observeEnabled() ? 'execute' : 'materialization'} />
-                )
-              }
+              icon={loading ? <Spinner purpose="body-text" /> : <Icon name="execute" />}
             >
               {firstOption.label}
             </Button>
