+++ js_modules/dagster-ui/packages/ui-core/src/assets/auto-materialization/AutomaterializationRoot.tsx
@@ -1,13 +1,11 @@
 import {Box, Page, PageHeader, Subtitle1} from '@dagster-io/ui-components';
 import {Redirect} from 'react-router-dom';
-import {observeEnabled} from 'shared/app/observeEnabled.oss';
 
 import {GlobalAutomaterializationContent} from './GlobalAutomaterializationContent';
 import {assertUnreachable} from '../../app/Util';
 import {useTrackPageView} from '../../app/analytics';
 import {AutomationTabs} from '../../automation/AutomationTabs';
 import {useDocumentTitle} from '../../hooks/useDocumentTitle';
-import {OverviewPageHeader} from '../../overview/OverviewPageHeader';
 import {useAutoMaterializeSensorFlag} from '../AutoMaterializeSensorFlag';
 
 // Determine whether the user is flagged to see automaterialize policies as
@@ -30,19 +28,12 @@ export const AutomaterializationRoot = () => {
 const GlobalAutomaterializationRoot = () => {
   useTrackPageView();
   useDocumentTitle('Automations | Auto-materialize');
-  const showTabs = observeEnabled();
   return (
     <Page>
-      {showTabs ? (
-        <>
-          <PageHeader title={<Subtitle1>Automation</Subtitle1>} />
-          <Box padding={{horizontal: 24}} border="bottom">
-            <AutomationTabs tab="global-amp" />
-          </Box>
-        </>
-      ) : (
-        <OverviewPageHeader tab="amp" />
-      )}
+      <PageHeader title={<Subtitle1>Automation</Subtitle1>} />
+      <Box padding={{horizontal: 24}} border="bottom">
+        <AutomationTabs tab="global-amp" />
+      </Box>
       <GlobalAutomaterializationContent />
     </Page>
   );
