--- js_modules/dagster-ui/packages/ui-core/src/home/HomeDarkLaunch.oss.tsx
@@ -1,2 +1,0 @@
-// OSS stub for component to perform dark launch queries for Dagster+ home.
-export const HomeDarkLaunch = () => <></>;
