--- js_modules/dagster-ui/packages/ui-core/src/nav/LeftNavItemType.tsx
@@ -1,18 +1,0 @@
-import {IconName} from '@dagster-io/ui-components';
-
-import {
-  WorkspaceRepositorySchedule,
-  WorkspaceRepositorySensor,
-} from '../workspace/WorkspaceContext/WorkspaceContext';
-import {RepoAddress} from '../workspace/types';
-
-export type LeftNavItemType = {
-  name: string;
-  isJob: boolean;
-  leftIcon: IconName | 'airflow';
-  label: React.ReactNode;
-  path: string;
-  repoAddress: RepoAddress;
-  schedules: WorkspaceRepositorySchedule[];
-  sensors: WorkspaceRepositorySensor[];
-};
