+++ js_modules/dagster-ui/packages/ui-core/src/overview/OverviewActivityRoot.tsx
@@ -3,8 +3,6 @@ import * as React from 'react';
 import {Redirect, Switch} from 'react-router-dom';
 
 import {OverviewAssetsRoot} from './OverviewAssetsRoot';
-import {OverviewPageHeader} from './OverviewPageHeader';
-import {OverviewTabs} from './OverviewTabs';
 import {OverviewTimelineRoot} from './OverviewTimelineRoot';
 import {Route} from '../app/Route';
 import {AssetFeatureContext} from '../assets/AssetFeatureContext';
@@ -14,13 +12,6 @@ import {ActivatableButton} from '../runs/ActivatableButton';
 
 export const OverviewActivityRoot = () => {
   useDocumentTitle('Overview | Activity');
-
-  const header = React.useCallback(
-    ({refreshState}: {refreshState: React.ComponentProps<typeof OverviewTabs>['refreshState']}) => (
-      <OverviewPageHeader tab="activity" refreshState={refreshState} />
-    ),
-    [],
-  );
 
   const [_defaultTab, setDefaultTab] = useStateWithStorage<'timeline' | 'assets'>(
     'overview-activity-tab',
@@ -57,11 +48,11 @@ export const OverviewActivityRoot = () => {
       <Switch>
         {!enableAssetHealthOverviewPreview && (
           <Route path="/overview/activity/assets">
-            <OverviewAssetsRoot Header={header} TabButton={tabButton} />
+            <OverviewAssetsRoot TabButton={tabButton} />
           </Route>
         )}
         <Route path="/overview/activity/timeline">
-          <OverviewTimelineRoot Header={header} TabButton={tabButton} />
+          <OverviewTimelineRoot />
         </Route>
         <Route
           path="*"
