--- js_modules/dagster-ui/packages/ui-core/src/ui/__stories__/SectionedLeftNav.stories.tsx
@@ -1,29 +1,0 @@
-import {LEFT_NAV_WIDTH} from '../../nav/LeftNav';
-import {
-  abcLocationOption,
-  fooBarOption,
-  loremIpsumOption,
-} from '../../nav/__fixtures__/LeftNavRepositorySection.fixtures';
-import {SectionedLeftNav} from '../SectionedLeftNav';
-
-// eslint-disable-next-line import/no-default-export
-export default {
-  title: 'SectionedLeftNav',
-  component: SectionedLeftNav,
-};
-
-export const Default = () => {
-  return (
-    <div style={{position: 'absolute', left: 0, top: 0, height: '100%', width: LEFT_NAV_WIDTH}}>
-      <SectionedLeftNav visibleRepos={[loremIpsumOption, fooBarOption, abcLocationOption]} />
-    </div>
-  );
-};
-
-export const SingleRepo = () => {
-  return (
-    <div style={{position: 'absolute', left: 0, top: 0, height: '100%', width: LEFT_NAV_WIDTH}}>
-      <SectionedLeftNav visibleRepos={[loremIpsumOption]} />
-    </div>
-  );
-};
